Add expression function for using MagDec within Layouts
It's useful to be able to display a map's magnetic declination alongside
the map itself on the layout.

This code creates a function that can be used in the Expression editor
to retrieve the center-point and crs of a given map within a Print
Layout, and return a double representing the magnetic declination at
that point.

This can then be used in a text area or similar on a print layout to
display that declination.